+++ lib/evil_events/core/events/manager/notifier.rb
@@ -5,7 +5,7 @@ class EvilEvents::Core::Events::Manager
   # @since 0.1.0
   module Notifier
     # @since 0.1.0
-    NotifierError = Class.new(StandardError)
+    NotifierError = Class.new(EvilEvents::Core::Error)
     # @since 0.1.0
     InconsistentEventClassError = Class.new(NotifierError)
 
@@ -14,6 +14,7 @@ class EvilEvents::Core::Events::Manager
       # @since 0.1.0
       extend Forwardable
 
+      # @since 0.1.0
       def_delegators :errors_stack, :<<, :empty?
 
       # @return [Concurrent::Array]