+++ lib/evil_events/core/events/serializers.rb
@@ -7,7 +7,7 @@ module EvilEvents::Core::Events
     extend EvilEvents::Shared::DependencyContainer::Mixin
 
     # @since 0.1.0
-    SerializersError = Class.new(StandardError)
+    SerializersError = Class.new(EvilEvents::Core::Error)
     # @since 0.1.0
     SerializationError = Class.new(SerializersError)
     # @since 0.1.0