+++ spec/units/evil_events/core/events/event_factory_spec.rb
@@ -161,6 +161,24 @@ describe EvilEvents::Core::Events::EventFactory, :stub_event_system do
         expect(manager_registry).to include(concrete_event_manager)
         expect(manager_registry.size).to eq(1)
       end
+
+      specify 'any other errors in event definition doesnt affects manager registry' do
+        event_type = gen_str
+
+        # create event class with errors
+        expect do
+          begin
+            described_class.create_class(event_type) { raise ZeroDivisionError }
+          rescue ZeroDivisionError
+            nil
+          end
+        end.not_to(change { manager_registry.size })
+
+        # create event class without errors
+        expect do
+          described_class.create_class(event_type)
+        end.to(change { manager_registry.size }.by(1))
+      end
     end
   end